move checks
